% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{capm}
\alias{capm}
\title{Expected return of a stock based on capm}
\usage{
capm(rf, b, mp)
}
\arguments{
\item{rf}{the risk free rate of return}

\item{b}{beta for the interested stock}

\item{mp}{the market premium rate}
}
\value{
capm expected return
}
\description{
Provide the risk-free rate of return, beta, and market premium to calculate the expected market return
}
