% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{capmValue}
\alias{capmValue}
\title{Value of a stock based on capm analysis}
\usage{
capmValue(capm, expR)
}
\arguments{
\item{capm}{the return according to capm}

\item{expR}{the expected return for the stock}
}
\value{
the stock's position
}
\description{
Provide the expected return and the capm return to determine the value of a stock using fundamental analysis
}
