% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finecn.R
\docType{package}
\name{finecn-package}
\title{finecn: Carries out basic finance based calculations and performs simple analysis}
\description{
This package contains functions that can be used to carry out common calculations used in basic finance.
  Additionally, the package contains the ability to perform funadmental analysis of stock returns using CAPM. 
  Technical analysis is also avaliable to find trends in stock price using moving averages.
}
\details{
Financial formulas and minor analysis
}
\author{
\strong{Maintainer}: Grant Smith \email{grant.oferrell.smith@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
