% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fv}
\alias{fv}
\title{Future Value of a cash flow}
\usage{
fv(pv, i, n)
}
\arguments{
\item{pv}{the present value of a cash flow}

\item{i}{the annual interest rate}

\item{n}{the number of periods}
}
\value{
Future Value
}
\description{
Provide the present value, interest rate, and number of periods
}
