% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fvannuity}
\alias{fvannuity}
\alias{pvannuity}
\title{Future Value of an ordinary annuity}
\usage{
fvannuity(cf, n, m, i, p = 1)

pvannuity(cf, i, n)
}
\arguments{
\item{cf}{the present value of a cash flow}

\item{n}{the number of years}

\item{m}{the number of periods in a year}

\item{i}{the annual interest rate}

\item{p}{an optional parameter to divide the number of}
}
\value{
Future Value

Present Value
}
\description{
Provide the cash flow, interest rate, and number of periods

Provide the cash flow, interest rate, and number of periods
}
