% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geom_Bar}
\alias{geom_Bar}
\title{Expected return of a stock based on the geometric mean}
\usage{
geom_Bar(x)
}
\arguments{
\item{x}{a vector containing the daily percent changes in price for a stock}
}
\value{
expected return of a stock
}
\description{
Provide the percentage change in the price of a stock to calculate the expected geometric return
}
