% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finecn.R
\name{mabar}
\alias{mabar}
\title{Calculate the moving average for a set of prices}
\usage{
mabar(prices, y)
}
\arguments{
\item{prices}{a vector of stock prices}

\item{y}{the interval used}
}
\value{
the nth-day moving average
}
\description{
Provide a vector of stock prices and the desired interval to get the nth day moving average
}
\examples{
prices<-round(rnorm(50,10.00),2)
mabar(prices,20)

}
