% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pv}
\alias{pv}
\title{Present Value of a cash flow}
\usage{
pv(fv, i, n)
}
\arguments{
\item{fv}{the future value of a cash flow}

\item{i}{the annual interest rate}

\item{n}{the number of periods}
}
\value{
Present Value
}
\description{
Provide the future value, interest rate, and number of periods to get the present value needed
}
